+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: spook <spook@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/11 12:14:25 by lvon-war          #+#    #+#             */
-/*   Updated: 2024/05/11 17:42:26 by spook            ###   ########.fr       */
+/*   Updated: 2024/05/12 03:51:28 by spook            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,5 @@ void		error_handler(char *txt, int code);
 t_color		int_to_color(int color);
 float       delta(float a, float b);
 t_color		pixel_plottin(t_img img, int x, int y);
-float		degtorad(int deg);
+float		degtorad(float deg);
 #endif /*!UTILS_H */
